feat: search an index from a file, on every platform that can
The clustered index existed for one property a graph cannot have — contiguity —
and the whole argument for it ended at a source that read from the same array
the index had been loaded from. The mechanism was there; the memory was not
reclaimed. This is the half that was missing.

## The seam, in kromus-core

ByteSource replaces VectorBlocks: three methods — size, read(offset, length,
into, at) and close() — over a whole container rather than a vector region.
ContainerReader now reads from one as well as from an array, through a single
header parser rather than two, because a second parser that drifted would accept
a file the first rejects. Given an array it still hands out sections as windows
and copies nothing; given a source it reads the header and nothing else until a
section is asked for.

openIvfIndex and openFlatIndex replace viewIvfIndex/viewFlatIndex. The old shape
took the whole file as a ByteArray *and* a place to read vectors from, which
meant you had to load the bulk into the heap in order to avoid loading it into
the heap. The new one takes a source and slices the vector section out of it.

KromusFormatException's constructor is public now: a source written outside the
library has to be able to report the same failures the library's own do.

What that costs is measured on the access pattern rather than argued. Opening a
600-entry index reads less than its vector section occupies; ten queries at
nprobe=3 together read less than the file. The assertions are against the size
of the vector region, so they stay meaningful when the corpus changes.

## The platform half, in kromus-files

A new optional module, separate because kromus-core is common code and nothing
else — four interop surfaces do not belong in an artifact whose whole claim is
that it has none. The seam stays in core, so a source of your own needs only
core.

  FileByteSource       JVM, Android          positional FileChannel reads, so one
                                             open file serves every searching thread
  FileByteSource       iOS, macOS, Linux,    fopen/fseek/fread; UnsafeNumber because
                       Windows               a C long is 64-bit on Unix and 32-bit
                                             on Windows
  NodeFileByteSource   Node (JS and Wasm)    fs.readSync
  OpfsByteSource       browser (JS and Wasm) a synchronous OPFS handle — worker only

openRange opens an index packed inside a larger file, which is the shape an
Android asset arrives in. Node's fs is reached through eval so that a bundler
building for the browser cannot see the import and fail on it.

The browser's only synchronous file read is OPFS's createSyncAccessHandle, and it
exists only inside a Web Worker — a constraint of the platform. OpfsByteSource
takes an already-open handle rather than opening one, so everything asynchronous
stays with the caller and ByteSource can stay synchronous.

## What is tested, and what is not

One shared contract, checked against every source: read the same bytes an array
does at every offset, honour the destination offset, refuse ranges outside
itself, refuse reads after closing, and loop rather than trust a handle that
reads short. That last one is the point of the exercise — a partly-filled buffer
is read as vector data, nothing throws, and the query comes back with neighbours
that merely look right.

The OPFS pair is tested against a stand-in handle that reproduces the browser's
contract, short reads included. The binding to the browser's own object is not
exercised by CI: a test runner drives the main thread, and a real handle cannot
be created there. Said plainly in the readme rather than left to be discovered.

## Release bookkeeping

0.16.0 was written into the changelog but never tagged or published, and its
format is superseded by the sectioned one here — publishing it would have handed
out a layout we already knew we were replacing. Its entries are folded into this
release, and the migration note now targets 0.15.0, which is what is actually on
Maven Central.

792 tests across JVM, linuxX64, JS/node and Wasm.

Closes #30.

+++ kromus-core/src/commonMain/kotlin/IvfSearcher.kt
@@ -88,7 +88,7 @@ public class IvfSearcher<K> internal constructor(
                 }
             }
         }
-        return top.toSortedList().map { SearchResult(it.key, index.similarity((-it.score).toFloat())) }
+        return top.toSortedList().map { SearchResult(it.key, index.similarity(-it.score)) }
     }
 
     /** False when this stored copy's entry has already been offered by an earlier list. */
